+++ src/Pim/Bundle/CatalogBundle/AttributeType/FileType.php
@@ -3,7 +3,6 @@
 namespace Pim\Bundle\CatalogBundle\AttributeType;
 
 use Pim\Bundle\FlexibleEntityBundle\Model\AbstractAttribute;
-use Pim\Bundle\FlexibleEntityBundle\AttributeType\FileType as FlexFileType;
 
 /**
  * File attribute type
@@ -12,7 +11,7 @@ use Pim\Bundle\FlexibleEntityBundle\AttributeType\FileType as FlexFileType;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class FileType extends FlexFileType
+class FileType extends AbstractAttributeType
 {
     /**
      * {@inheritdoc}